Improve cleaning in nightly_build.sh

+++ nightly_build.sh
@@ -7,14 +7,13 @@ rm -f nightly_build.log
 
 # handle clean here to avoid removing the log
 
-if [ "$1" = clean ] || [ "$1" = "--clean" ]
+if [ "$1" != no-clean ] && [ "$1" != "--no-clean" ]
 then
-	shift
 	echo "Cleaning..."
 	git clean -dfx
 fi
 
 (
  echo "Starting nightly_build $* at $(date) in $(pwd)"
- ./scripts/makedist.sh "$@"
+ ./scripts/makedist.sh --no-clean "$@"
 ) 2>&1 | tee nightly_build.log
